mv polling bookkeeping into own module
